make sure the tests has proper namespace
